+++ topi/python/topi/vision/__init__.py
@@ -5,3 +5,4 @@ from __future__ import absolute_import as _abs
 from . import yolo2, ssd
 from .shortcut import *
 from .reorg import *
+from .nms import *
